libfstt: Change faceLength type from int to size_t
The variable stores and is used to compare string sizes, which are
all positive.

+++ libfstt/ttf.h
@@ -200,7 +200,7 @@ struct Point {
 struct FontInfo {
 	uint16_t firstChar, lastChar;
 	uint8_t panose[10];
-	int	faceLength;
+	size_t faceLength;
 	char	faceName[32];
 };
